refactor(012): extract session-cache from session.ts (US3)
Move the pure cache-building helpers (buildCacheFromRecord, buildPageCache,
registerPageQuestions, updateCacheWithAnswer) out of session.ts into
session-cache.ts so SessionService retains only session lifecycle. Re-point
storage-service and quiz-table imports; relocate the cache unit tests to
tests/unit/session-cache.test.ts. session.ts is now under 300 lines.
